Update patch set 3
Patch Set 3:

Adam, I think it might be reasonable to add this as an option but ideally it should be done in coordinate with other providers of this API (OpenSSL and/or LibreSSL). Also, whether or not it makes sense to add the option depends on whether Google would actually use it; if not, then nobody actually would. In *ring*, we don't even use this code; somebody is working on doing all the PSS logic in Rust. 

In general, the response from other people has been similar to yours; people are generally skeptical that this is a good idea. I'll probably put this idea on ice for a while and hope that I get some more positive and more definitive results in favor of FDH being as good as probabilistic signatures.

I did update this PR to add some counter-claims to the original claims I cited, which people clued me into in the interim.

Thanks for taking the time to look at this.

Patch-set: 3
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0