Update patch set 3
Patch Set 3: Commit-Queue+1

Piotr: Sadly, we found a second project which depends on this (currently in a way that isn't broken by 9498e74, but in a bad way). We can probably get them fixed too, but with two data points, it seems we ought to preserve this behavior.

Adam: How's this look? I extended the test to run at all versions and check for resumption. Actually it found that the SSL 3.0 code was missing a call too. I'm unhappy at how many places this is peppered though. :-/

Patch-set: 3
Label: Commit-Queue=+1