Update patch set 2
Patch Set 2:

> (Although they might be wired into Android somewhere.)

They are, which is largely why I've been stalling on https://crbug.com/boringssl/2. :-)

https://android.googlesource.com/platform/external/conscrypt/+/master/src/main/native/org_conscrypt_NativeCrypto.cpp#6950

But I don't think anything will break if we change it, short of the usual ancient Java clients breaking if we go beyond 1024-bit. But if we're going to be putting much effort there, maybe we should ask them to just take it out altogether. There's no way DHE is performant enough on a tiny little phone anyway.

Patch-set: 2