Draft: Switch OPENSSL_VERSION_NUMBER to 1.1.0.
Work in progress. This is sufficient to get Android building, but we'll
need a few more functions both for completeness and compatibility with
other projects.

(Turns out adding accessors has to be atomic with swapping the version
because C refuses to have function declarations and static functions
with the same name.)

Change-Id: I5eeb27448a6db4c25c244afac37f9604d9608a76