Draft: Switch OPENSSL_VERSION_NUMBER to 1.1.0.
Work in progress.

This atomicly adds accessors for OpenSSL 1.1.0 and switches
OPENSSL_VERSION_NUMBER. This must be atomic because downstream projects
will often provide their own compatibility versions of accessors based
on a preprocessor check. That means OpenSSL versions prior to 1.1.0 must
*not* provide the corresponding functions or the build will fail.

This CL includes enough accessors that projects which have already
ported 1.1.0 should build. At the same time, it does not remove any
1.0.2 APIs, so we will not require that all projects do. The exception
is APIs which changed in 1.1.0 but did not change the function
signature. Those are breaking changes we'll need to get fixed sooner.
Specifically:

- SSL_CTX_sess_set_get_cb is now const-correct.

- X509_get0_signature is now const-correct.

Bug: 91
Change-Id: I5eeb27448a6db4c25c244afac37f9604d9608a76