Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Yeah, but then we'll have some shared state on hs, some TLS 1.2 specific state and some TLS 1.3 specific state. I was thinking of having a flag to clear the state we won't need after version negotiation, instead of freeing each bit separately.
> 
> ssl_handshake_clear(SSL_HANDSHAKE *hs, version_t version);

Do you anticipate there being much (any?) state beyond this? It's only stuff needed in creating the ClientHello. The rest is just going to be random booleans and such. I'd prefer leaving it simple for now and we can generalize later if we need it.

Also nothing bad happens if we don't drop it early. The nice thing about the separate SSL_HANDSHAKE structure is that it's fairly short-lived, during a time period where we'll already be managing even larger temporaries (namely the certificate message). This was mostly a "may as well, I guess" sort of thing.

Patch-set: 1