Update patch set 11
Patch Set 11:

Consider the case where you are building BoringSSL on an old version of Linux that doesn't have the Linux headers that define the getrandom stuff. Then, IIUC, `getrandom` won't be used by the generated executable. But, really, the configuration of the machine building BoringSSL shouldn't decide this; it should be decided by the configuration of the machine where the executable is run.

On the other hand, if you rely on your own definitions of the syscall numbers then there is some small chance that they might be wrong.

So, I recommend that you invert the logic a bit: Always enable getarndom support, using your own definitions of the syscall numbers. If the Linux headers also define `SYS_getrandom` then verify that it matches what you would use. But, if the Linux headers don't define `SYS_getrandom` then just keep going.

Otherwise, it is important to document in BUILDING.md and/or PORTING.md that one must use newer (newer than is typically supplied by Linux distros) Linux headers in order to get getrandom support.

Patch-set: 11