std::default_delete all the things.
This replaces ScopedFOO with std::default_delete for all straightforward
heap-allocated cases.

The declaration of default_delete lives in base.h.  This is something we
have to remember to do, but if we do it, it ensures that it is an error
to not have default_delete available in any context where a unique_ptr
of the type in question goes out of scope.  (The definition of
default_delete lives inline in the header file that declares the freeing
function.)

The C++ code is wrapped in `#if defined(__cplusplus)' and `extern
"C++"', the latter in case there is already an enclosing `extern
"C"'.  (Whyyyy?)

Change-Id: Ife6a5de1b2528a25cb493452fb498e6ef7b63b03