Remove the BORINGSSL_ANDROID_SYSTEM P-521 special-case.
This dates to
https://android.googlesource.com/platform/external/boringssl/+/62d05888d1cf178d900b54e7e035928abea512b1
which intended to be removed in a later Android release once X25519 was
added. That has since happened.

This intentionally leaves the P-521 hooked up for now. Detaching it
completely is a more aggressive change (since it's slightly tied up with
SHA-512) that should wait until removing ECDSA+SHA512 has stuck in Chrome.

Change-Id: I04553c3eddf33a13b6e3e9a6e7ac4c4725676cb0