Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > This might be best in decrepit/bio/bio_decrepit.c (new file) so
 > that we notice if anything tries to use it. Does libdecrepit work
 > for you?

Hmmm, this is IMO a useful function and it's used by many projects, so I'm not sure I would consider it deprecated. It is not much different from, say, BIO_reset or BIO_flush.

The downside of moving this in a separate library is that I would need to change the NGINX/OpenResty build system to check for and use this other thing, which is something I'd rather not do.

Why would you consider this "decrepit"? The BIO_pair incompatibility?

Patch-set: 1
Reviewer: Gerrit User 7060 <7060@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0