Update patch set 4
Patch Set 4:

(8 comments)

Did a first pass of this mostly thinking about tests. When adding new features like this, it is important that *in the same CL* one thinks about all the cases that are worth testing and add tests there. Otherwise we'll forget what was and wasn't covered.

Patch-set: 4