Update patch set 1
Patch Set 1:

> What's the motivation here? Upstream makes them mutually-exclusive,
 > so it'd be a departure from their behavior. This is also how most
 > settings work (although they're usually not implemented this way).
 > If you set something on the SSL, the setting on the SSL_CTX is
 > ignored.

Ah, then this is probably just a pattern I wasn't aware of. 

I tried setting the callback on the ctx and was surprised to find out it didn't work because some other code had already registered a callback on the ssl. I couldn't find a mention of this in the docs.

Patch-set: 1
Reviewer: Gerrit User 7160 <7160@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0