Update patch set 2
Patch Set 2:

> (1 comment)

It's much easier to reject duplicates in a sorted list, but if the order of the list is already meaningful then we can't sort it.

But how about just enforcing that they match the order of the supported_groups now and make it a MUST in the spec later?

Patch-set: 2
Label: Code-Review=0