Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > (1 comment)
> 
> It's much easier to reject duplicates in a sorted list, but if the order of the list is already meaningful then we can't sort it.
> 
> But how about just enforcing that they match the order of the supported_groups now and make it a MUST in the spec later?

I briefly made it a MUST in the spec, but we backed it out. It turns out the whole notion of key shares being the client pref order is bust because a HelloRetryRequest + second ClientHello makes the unpredicted key share go at the end. It also doesn't prevent duplicates because we don't check for duplicates in the group list either.

I suggest we leave this as-is for now and we'll sort out key share's ordering problems asynchronously.

Patch-set: 2
Label: Code-Review=0