Update patch set 1
Patch Set 1:

We don't seem to have a CLA on file for martin.thomson@gmail.com. (https://cla.developers.google.com/about)

Happy to remove P-224 if it's in the way. We don't have it hooked up to the TLS stack, so there's no need for it to work in BoGo. Though are you sure Go 1.7 lacks it? Our CI uses 1.7, and it's fine.
https://golang.org/pkg/crypto/elliptic/#P224

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0