Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm just squashing the compilation warning so this is probably not that great now that I think on it.  The best solution is to remove your dependency on readdir_r().

Well, it's in decrepit for a reason. :-) But, yeah, I'll go switch that to readdir. The sordid mess that is readdir vs readdir_r is sad:
https://www.gnu.org/software/libc/manual/html_node/Reading_002fClosing-Directory.html

(PS: I haven't looked at your CL yet since we don't seem to have a CLA on file for the email you used. That or I don't know how to use the tool. But it sounds like it won't be necessary here.)

Patch-set: 1