Add a temporary knob to disable RSA-PSS.
Older versions of Erlang's TLS stack are intolerant to unknown signature
algorithms. Thankfully, this bug isn't widespread, but some consumer
needs to be able to talk with one such broken server. To avoid
interruption, add a temporary knob to let them turn RSA-PSS off, but
this needs to be resolved ASAP.

Change-Id: If3a5662b419513738c189cc94e513926cea0efe9