Update patch set 14
Patch Set 14: Code-Review+1 Commit-Queue+1

(7 comments)

Thanks! I think this is good. Some minor comments on the Go side (mostly one more test that would be nice to have).

I'm not sure how we want to land this... WDYT? It's a breaking protocol change that doesn't bump the version. Since this change is really part of the cipher suites one too, perhaps we should just land them in tandem once the cipher suites one is good?

Or we could decide we don't really care and land this as it is.

Patch-set: 14
Label: Code-Review=+1
Label: Commit-Queue=+1