Define SSL_CTRL_SET_CURVES_LIST
This is used by NGINX to enable the SSL_CTX_set1_curves_list() API which
in turn makes it possible to use X25519.

If that is not defined NGINX falls back to using EC_KEY_new_by_curve_name()
which doesn't know about X25519 (which seems a bug on its own).

Change-Id: I890a801484db3a2a6f4f7bdb93f8cd3490f5924e