Update patch set 1
Patch Set 1: Commit-Queue+1

I'm going to try to find some time either this weekend or next week to figure out what this code is doing. In the meantime, I wanted to sort out the more mechanical diff-munging stuff before trying to digest it all. Does this look insane to either of you?

Brian, do I understand the thread between you and Andy correctly that various operations weren't fully reducing to [0, P) but now they all are? (Andy keeps talking about "preserving" the property but it's not clear to me what the promises are if the input wasn't reduced...) In that case, should the "where res is not fully reduced" comments you added to p256-x86_64.c be removed, or are they still correct?

It's late over on this coast, so I'll deal with the is_infinity one after this. Am I correct that they're independent?

Patch-set: 1