Update patch set 1
Patch Set 1:

> Brian, do I understand the thread between you and Andy correctly that various operations weren't fully reducing to [0, P) but now they all are?

That's the intent.

> (Andy keeps talking about "preserving" the property but it's not clear to me what the promises are if the input wasn't reduced...)

Simplifying a bit, the output will be fully reduced if the input is fully reduced.

> In that case, should the "where res is not fully reduced" comments you added to p256-x86_64.c be removed, or are they still correct?

There are parts of the p256-x86_64.c code that would be wrong if the point addition/doubling functions return coordinates that aren't reduced. The point of Andy's patch is to ensure that all the outputs are fully reduced. Assuming that is true, those comments are now misleading. Also the comment I added in the inversion code that the from_mont conversion must be last is (IIUC) now misleading.

> It's late over on this coast, so I'll deal with the is_infinity one after this. Am I correct that they're independent?

There are/were multiple potential issues with points at infinity in the code. I believe some of them are resolved by this patch, and some of them are resolved with the other patch(es).

Patch-set: 1
Label: Code-Review=0