Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Having read through the linked thread:
>   a) I'm glad Brian exists.
>   b) The world need formal verification tools for this stuff.
> 
> Here's a plan:
> 
> Let's land this change and also pick up the point at infinity changes.
> 
> Let's write a test function that drives the basic operations and ensures that the invariants that we believe are now true, are actually true. (This test function will need to be in the production code because a shared-library test won't be able to access all the functions needed. That should be fine as the linker can discard it, esp if it's in a separate file.)
> 
> Then write tests for each of the infinity cases that Brian identified.

I was thinking we'd put the functions in some internal header and the test would #ifdef itself out if built as a shared library. Not great, but otherwise the linker can't drop the tests in shared library builds.

(I'll update this CL to fix the comments in p256-x86_64.c per Brian's comment.)

Patch-set: 1