Don't depend parse_clienthello hooks on the resumption decision.
This is in preparation for parsing ClientHello extensions before
resumption. Which is in preparation for calling the legacy SNI callback
and cert_cb before resumption. Which is in preparation for resolving the
cipher suite before resumption and only resuming if the negotiated
cipher and the one in the session match.

Change-Id: I9bba0f84d975255bfbf4fc65e7df4369af1d5757