Parse ClientHello extensions before deciding on resumption.
This simplifies a little code around extended master secret. The
extensions that are more tightly integrated with the handshake are still
processed inline for now. It does, however, require an extra state in
1.2 so the asynchronous session callback does not cause extensions to be
processed twice. Tweak a test enforce this.

This and a follow-up to move cert_cb before resumption are done in
preparation for resolving the cipher suite before resumption and only
resumption on match.

Note this has caller-visible effects:

- The legacy SNI callback happens before resumption.

- The ALPN callback happens before resumption.

- Custom extension ClientHello parsing callbacks also cannot depend on
  resumption state.

- The DoS protection callback now runs after all the extension callbacks
  as it is documented to be called after the resumption decision.

Change-Id: I1281a3b61789b95c370314aaed4f04c1babbc65f