Update patch set 1
Patch Set 1: Code-Review-1

We don't want to build with OPENSSL_NO_ASM on Android. There is ARM assembly and we need to test it during development.

That the assembly doesn't work with the NDK toolchain file is a bug in the NDK. See here:
https://crbug.com/boringssl/109

I've already fixed the bug in the NDK itself. I'm planning on just waiting for them to cut a new build with the fix in it and, until then, just stick with the android-cmake thing.

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=-1