Update patch set 4
Patch Set 4:

> I'd actually be pretty happy saying that supported_versions means you can't go below 1.2 at either client or server.

I agree that that is a good idea.

> And I don't mind sending a marginally smaller ClientHello.

Exactly.

> Although we probably should get the spec to decide on it one way or another and make client/server agree.

I submitted this PR so as to inform that discussion. I think it would be useful for BoringSSL to make this change now prior to the interop testing at IETF to show that it isn't problematic.

I updated the comment w.r.t. your idea about fingerprinting. Thanks!

Patch-set: 4
Label: Code-Review=0