Update patch set 18
Patch Set 18:

> We don't, generally, mark ignored return values with (void).

OK, sorry for the noise. I didn't realize that. (IMO, such marking is a good idea because it makes it obvious that the lack of checking is intentional.)

I also agree that trying to do anything with the return value in this test is probably not worth the effort.

Thanks!

Patch-set: 18